requirements.txt
google-adk
yfinance
psutil
litellm
google-generativeai
python-dotenv
pydantic
fastapi